addressing comments
